+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Refactoring deprecated classes. see https://github.com/techdivision/import-cli-simple/blob/master/UPGRADE-4.0.0.md
 * Add techdivision/import-product-variant#22
 * PAC-294: Integration strict mode
+* PAC-541: Update composer with php Version ">=^7.3"
 
 # Version 24.0.1